Move VotacoesSemelhantes para a pasta de componentes comuns
